chore: major refactorings
This change includes:
- pathetic-mapper is no more and was replaced with a specific module pathetic-bukkit
- pathetic-model is now pathetic-engine
- pathetic-engine is now independent of the bukkit api
- the required shutdown process was replaced with a shutdown hook
- the PathfinderFactory is now part of the pathetic-api and has to be instantiated with the respective PathfinderInitializer
