+++ pathetic-engine/src/main/java/de/metaphoriker/pathetic/Pathetic.java
@@ -1,102 +1,30 @@
 package de.metaphoriker.pathetic;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.HashSet;
-import java.util.Properties;
-import java.util.Set;
-
-import com.google.common.eventbus.Subscribe;
-import de.metaphoriker.pathetic.api.event.EventPublisher;
-import de.metaphoriker.pathetic.api.event.PathingStartFindEvent;
+import de.metaphoriker.pathetic.util.ErrorLogger;
 import lombok.Getter;
 import lombok.experimental.UtilityClass;
-import lombok.extern.slf4j.Slf4j;
-import org.bukkit.Bukkit;
-import org.bukkit.plugin.java.JavaPlugin;
-import de.metaphoriker.pathetic.bukkit.listeners.ChunkInvalidateListener;
-import de.metaphoriker.pathetic.util.BukkitVersionUtil;
-import de.metaphoriker.pathetic.util.ErrorLogger;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
 
 @UtilityClass
-@Slf4j
 public class Pathetic {
 
   private static final String PROPERTIES_FILE = "pathetic.properties";
 
-  private static final Set<Runnable> SHUTDOWN_LISTENERS = new HashSet<>();
+  @Getter
+  private static String engineVersion;
 
-  private static JavaPlugin instance;
-  @Getter private static String modelVersion;
-
-  /**
-   * @throws IllegalStateException If an attempt is made to initialize more than 1 time
-   */
-  public static void initialize(JavaPlugin javaPlugin) {
-
-    if (instance != null) throw ErrorLogger.logFatalError("Can't be initialized twice");
-
-    instance = javaPlugin;
-    Bukkit.getPluginManager().registerEvents(new ChunkInvalidateListener(), javaPlugin);
-
-    registerPathingStartListener();
-    loadModelVersion();
-
-    if (BukkitVersionUtil.getVersion().isUnder(16, 0)
-        || BukkitVersionUtil.getVersion().isEqual(BukkitVersionUtil.Version.of(16, 0))) {
-      log.warn(
-          "Pathetic is currently running in a version older than or equal to 1.16. "
-              + "Some functionalities might not be accessible, such as accessing the BlockState of blocks.");
-    }
-
-    log.debug("Pathetic v{} initialized", modelVersion);
-  }
-
-  public static void shutdown() {
-    SHUTDOWN_LISTENERS.forEach(Runnable::run);
-    SHUTDOWN_LISTENERS.clear();
-
-    instance = null;
-    log.debug("Pathetic shutdown");
-  }
-
-  public static boolean isInitialized() {
-    return instance != null;
-  }
-
-  public static JavaPlugin getPluginInstance() {
-    return instance;
-  }
-
-  public static void addShutdownListener(Runnable listener) {
-    SHUTDOWN_LISTENERS.add(listener);
-  }
-
-  private static void registerPathingStartListener() {
-    EventPublisher.registerListener(
-        new Object() {
-          @Subscribe
-          public void onPathingStart(PathingStartFindEvent event) {
-            pushToBStatsIfActivated(event);
-          }
-
-          private void pushToBStatsIfActivated(PathingStartFindEvent event) {
-            if (event.getPathfinderConfiguration().isBStats()) {
-              BStatsHandler.increasePathCount();
-            }
-          }
-        });
-  }
-
-  private static void loadModelVersion() {
+  public static void loadEngineVersion() {
     try (InputStream inputStream =
-        Pathetic.class.getClassLoader().getResourceAsStream(PROPERTIES_FILE)) {
+           Pathetic.class.getClassLoader().getResourceAsStream(PROPERTIES_FILE)) {
       Properties properties = new Properties();
       properties.load(inputStream);
 
-      modelVersion = properties.getProperty("model.version");
+      engineVersion = properties.getProperty("engine.version");
     } catch (IOException e) {
-      throw ErrorLogger.logFatalError("Error loading model version", e);
+      throw ErrorLogger.logFatalError("Error loading engine version", e);
     }
   }
 }
