+++ pathetic-engine/src/main/resources/pathetic.properties
@@ -1,1 +1,1 @@
-model.version=${project.version}
+engine.version=${project.version}
