+++ pathetic-example/pom.xml
@@ -72,7 +72,7 @@
         </dependency>
         <dependency>
             <groupId>de.metaphoriker</groupId>
-            <artifactId>pathetic-mapping</artifactId>
+            <artifactId>pathetic-bukkit</artifactId>
             <version>3.3.1</version>
         </dependency>
     </dependencies>
