+++ pathetic-example/src/main/java/de/metaphoriker/pathetic/example/PatheticPlugin.java
@@ -1,10 +1,12 @@
 package de.metaphoriker.pathetic.example;
 
 import de.metaphoriker.pathetic.api.pathing.Pathfinder;
+import de.metaphoriker.pathetic.api.pathing.PathfinderFactory;
 import de.metaphoriker.pathetic.api.pathing.configuration.PathfinderConfiguration;
+import de.metaphoriker.pathetic.bukkit.PatheticBukkit;
+import de.metaphoriker.pathetic.bukkit.factory.BukkitPathfinderFactory;
+import de.metaphoriker.pathetic.bukkit.factory.BukkitPathfinderInitializer;
 import de.metaphoriker.pathetic.example.command.PatheticCommand;
-import de.metaphoriker.pathetic.mapping.PatheticFacade;
-import de.metaphoriker.pathetic.mapping.PathfinderFactory;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class PatheticPlugin extends JavaPlugin {
@@ -13,12 +15,16 @@ public final class PatheticPlugin extends JavaPlugin {
   @Override
   public void onEnable() {
 
-    // Initialize the PatheticMapper with this plugin instance
-    PatheticFacade.initialize(this);
+    // Initialize Pathetic with this plugin instance
+    PatheticBukkit.initialize(this);
+
+    // Create the respective Pathfinder Factory with the respective initializer
+    PathfinderFactory<BukkitPathfinderInitializer> factory =
+        new BukkitPathfinderFactory(new BukkitPathfinderInitializer());
 
     // Create a new Pathfinder instance with a custom configuration
     Pathfinder reusablePathfinder =
-        PathfinderFactory.createPathfinder( // Use the factory to create a new pathfinder instance
+        factory.createPathfinder( // Use the factory to create a new pathfinder instance
             PathfinderConfiguration.createConfiguration()
                 .withAllowingFallback(true) // Allow fallback strategies if the primary fails
                 .withLoadingChunks(true) // Allow chunks to be loaded during pathfinding
@@ -26,12 +32,5 @@ public final class PatheticPlugin extends JavaPlugin {
 
     // Register the command executor for the "pathetic" command
     getCommand("pathetic").setExecutor(new PatheticCommand(reusablePathfinder));
-  }
-
-  // Called when the plugin is disabled
-  @Override
-  public void onDisable() {
-    // Shutdown the PatheticMapper to clear any resources it holds
-    PatheticFacade.shutdown();
   }
 }
